\name{print.VarSel}
\alias{print.VarSel}
\alias{print.VarSelPerformance}
\alias{print}

\title{Print objects from SUPCAVS package}

\description{Print two kind of objects: one object obtained from variable selection procedure and the other, obtained from performance test procedure.}

\usage{
print(x, ...)
}

\arguments{
  \item{x}{The object can be of type VarSel (Variable Selection) obtained from pcavs or spcavs, or can be an object of type VarSelPerformance obtained from "modelPerformance" function.}
  \item{...}{Any other parameter that can be passed to default print function.}
}
