\name{supcavs-package}

\alias{supcavs-package}

\alias{supcavs}

\docType{package}

\title{Supervised and Unsupervised PCA Variable Selection}

\description{This package provide two ways of using some PCA criterion for variable selection. One way is to use unsupervised PCA to do the variable selection, as explained in Jolliffe (1972). The other way is to use the Supervised PCA technique, as explained by Bair et al. (2006). The work of Oliveira Jr. et al. (2012) explain how to lead with PCA and SPCA to implement supervised and unsupervised variable selection using PCA.

To start using this package, take a look at help of functions "spcavs" or "pcavs".
}

\details{
\tabular{ll}{
Package: \tab spcavs\cr
Type: \tab Package\cr
Version: \tab 1.2 \cr
Date: \tab 2013-07-05 \cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes \cr
}

}
\author{MSc. Juscelino Izidoro de Oliveira Jr., PhD. José Carlos Ferreira da Rocha and PhD. Adriel Ferreira da Fonseca

Maintainer: PhD. José Carlos Ferreira da Rocha <jrocha@uepg.br>
}

\references{Jolliffe, I. T. (1972) Discarding Variables in a Principal Component Analysis. I: Artificial Data. Journal of the Royal Statistical Society. Series C (Applied Statistics). 21, 160 - 173.

Bair, E.; Hastie, T.; Paul, D.; Tibshirani, R. (2006). Prediction by Supervised Principal Components. Journal of the American Statistical Association. 101, 119 - 137.

Oliveira Jr, J. I.; Rocha, J. C. F.; Fonseca, A. F. (2012). Variable selection in agricultural data mining: an approach based in PCA (In Brazillian Portuguese). Master Thesis of Applied Computing Course. Ponta Grossa State University.

}
